+++ stl/lib/stl.dart
@@ -8,4 +8,5 @@ export 'src/collections/deque.dart';
 export 'src/collections/stack.dart';
 export 'src/collections/forward_list.dart';
 export 'src/ranges/numberline.dart';
+export 'src/utility/pair.dart';